refactor(vm): split the sandbox record store from its phase vocabulary
`persistence.rs` held two things at once: what a durable record may say
about a sandbox generation, and how that record reaches the disk. R3's
lifecycle HSM is written against the first half, so it moves out on its
own.

`record/phase.rs` carries the phase enum and its legal edges, the record
with its transitions and redaction, and the provisioning classification;
`record/store.rs` keeps the generation-checked, durability-reporting file
store. `record.rs` re-exports both halves, so the only call-site change is
the module name (`super::persistence::X` -> `super::record::X`) in nine
files.

This is a pure move — 436 insertions / 386 deletions, every body
byte-identical to its old form. The only edits are visibility requotes:
`pub(super)` inside `persistence` meant "visible in `sandbox`", which from
a child of `record` is spelled `pub(in crate::sandbox)`; items only the
store needs (`RECORD_VERSION`, `SandboxRecord::{new, apply}`,
`classify_existing_provision`, `validate_record`) stay one level tighter.
rustfmt rewraps three signatures the longer token pushes past 100 columns.
