+++ virt/arcbox-vm/src/sandbox/record/store.rs
@@ -1,3 +1,11 @@
+//! The file-backed record store: how a [`super::phase::SandboxRecord`]
+//! reaches the disk, and what "durable" means for each write.
+//!
+//! Every mutation is generation-checked, serialized by a process-local
+//! mutex, and reports its own durability separately from its visibility
+//! ([`DurableCommit`]) — a rename that landed is still the record even when
+//! the directory fsync behind it could not be confirmed.
+
 use std::ffi::OsStr;
 use std::fs::{self, File};
 use std::os::unix::fs::PermissionsExt;
@@ -5,298 +13,32 @@ use std::path::{Path, PathBuf};
 use std::sync::{Mutex, MutexGuard};
 
 use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use tracing::warn;
 use uuid::Uuid;
 
 use arcbox_atomic_file::AtomicWriteError;
 
-use super::types::IdleAction;
-use super::{SandboxId, SandboxSpec, validate_id};
+use super::phase::{
+    ExistingProvision, ProvisionIntent, RECORD_VERSION, SandboxPhase, SandboxProvisionOutcome,
+    SandboxRecord, SandboxTransition, classify_existing_provision, validate_record,
+};
 use crate::error::{Result, VmmError};
+use crate::sandbox::types::IdleAction;
+use crate::sandbox::{SandboxSpec, validate_id};
 
-const RECORD_VERSION: u32 = 1;
 const RECORDS_DIR: &str = "sandbox-records";
-
-/// Durable control-plane phase for one sandbox generation.
-///
-/// Workload execution remains transient: a running workload keeps the durable
-/// phase at `Ready`, avoiding record writes on the execution hot path.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub(super) enum SandboxPhase {
-    Creating,
-    Starting,
-    Ready,
-    Stopping,
-    Stopped,
-    Failed,
-    Removing,
-    /// Pause in progress: checkpoint taken or being taken, runtime
-    /// resources not yet fully released.
-    Pausing,
-    /// Checkpointed with resources released; `pause_snapshot_id` names the
-    /// catalog entry a Resume restores from (CORE-21).
-    Paused,
-    /// Resume in progress: runtime resources are being re-created.
-    Resuming,
-}
-
-/// The stable result returned once a provisioning request has been accepted.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(super) struct SandboxProvisionOutcome {
-    pub(super) ip_address: String,
-}
-
-/// Versioned durable state for one sandbox generation.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub(super) struct SandboxRecord {
-    version: u32,
-    pub(super) id: SandboxId,
-    pub(super) generation: Uuid,
-    pub(super) request_key: String,
-    pub(super) effective_spec: SandboxSpec,
-    pub(super) phase: SandboxPhase,
-    pub(super) provision_outcome: Option<SandboxProvisionOutcome>,
-    pub(super) created_at: DateTime<Utc>,
-    pub(super) error: Option<String>,
-    /// Catalog id of the internal pause checkpoint. Set while the record is
-    /// `Paused`/`Resuming` so a Resume after an agent restart still finds
-    /// its snapshot; cleared when the sandbox is `Ready` again.
-    #[serde(default)]
-    pub(super) pause_snapshot_id: Option<String>,
-    /// When the sandbox reached `Paused` (None otherwise).
-    #[serde(default)]
-    pub(super) paused_at: Option<DateTime<Utc>>,
-    /// When the hard maximum lifetime fires (None = no limit). Seeded from
-    /// `effective_spec.ttl_seconds`; replaced by `SetLifecycle` (CORE-60).
-    /// Survives `redact_runtime_inputs` — unlike the seed seconds, the
-    /// deadline is durable lifecycle state, not a runtime input.
-    #[serde(default)]
-    pub(super) ttl_deadline: Option<DateTime<Utc>>,
-}
-
-/// Result of reserving a durable provisioning intent.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(super) enum ProvisionIntent {
-    Created(SandboxRecord),
-    Resume(SandboxRecord),
-    Replay(SandboxRecord),
-    Blocked(SandboxRecord),
-}
-
-/// Generation-checked lifecycle update.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(super) enum SandboxTransition {
-    Starting(SandboxProvisionOutcome),
-    ReadyWithOutcome(SandboxProvisionOutcome),
-    Ready,
-    Stopping,
-    Stopped,
-    Failed(String),
-    Removing,
-    Pausing,
-    Paused { snapshot_id: String },
-    Resuming,
-}
-
-impl SandboxTransition {
-    fn phase(&self) -> SandboxPhase {
-        match self {
-            Self::Starting(_) => SandboxPhase::Starting,
-            Self::ReadyWithOutcome(_) | Self::Ready => SandboxPhase::Ready,
-            Self::Stopping => SandboxPhase::Stopping,
-            Self::Stopped => SandboxPhase::Stopped,
-            Self::Failed(_) => SandboxPhase::Failed,
-            Self::Removing => SandboxPhase::Removing,
-            Self::Pausing => SandboxPhase::Pausing,
-            Self::Paused { .. } => SandboxPhase::Paused,
-            Self::Resuming => SandboxPhase::Resuming,
-        }
-    }
-}
-
-impl SandboxRecord {
-    fn new(id: &str, request_key: &str, effective_spec: SandboxSpec) -> Self {
-        let created_at = Utc::now();
-        let ttl_deadline = (effective_spec.ttl_seconds > 0)
-            .then(|| created_at + chrono::Duration::seconds(i64::from(effective_spec.ttl_seconds)));
-
-        Self {
-            version: RECORD_VERSION,
-            id: id.to_owned(),
-            generation: Uuid::new_v4(),
-            request_key: request_key.to_owned(),
-            effective_spec,
-            phase: SandboxPhase::Creating,
-            provision_outcome: None,
-            created_at,
-            error: None,
-            pause_snapshot_id: None,
-            paused_at: None,
-            ttl_deadline,
-        }
-    }
-
-    fn apply(&mut self, transition: SandboxTransition) -> Result<()> {
-        let next = transition.phase();
-        let atomic_ready = matches!(&transition, SandboxTransition::ReadyWithOutcome(_))
-            && self.phase == SandboxPhase::Creating;
-        if !atomic_ready && !self.phase.can_transition_to(next) {
-            return Err(VmmError::WrongState {
-                id: self.id.clone(),
-                expected: format!("a valid transition from {}", self.phase.as_str()),
-                actual: next.as_str().to_owned(),
-            });
-        }
-
-        match transition {
-            SandboxTransition::Starting(outcome) => {
-                if let Some(existing) = &self.provision_outcome
-                    && existing != &outcome
-                {
-                    return Err(VmmError::WrongState {
-                        id: self.id.clone(),
-                        expected: format!("provision outcome {existing:?}"),
-                        actual: format!("provision outcome {outcome:?}"),
-                    });
-                }
-                self.phase = SandboxPhase::Starting;
-                self.provision_outcome = Some(outcome);
-                self.error = None;
-            }
-            SandboxTransition::ReadyWithOutcome(outcome) => {
-                if let Some(existing) = &self.provision_outcome
-                    && existing != &outcome
-                {
-                    return Err(VmmError::WrongState {
-                        id: self.id.clone(),
-                        expected: format!("provision outcome {existing:?}"),
-                        actual: format!("provision outcome {outcome:?}"),
-                    });
-                }
-                self.phase = SandboxPhase::Ready;
-                self.provision_outcome = Some(outcome);
-                self.error = None;
-                self.redact_runtime_inputs();
-            }
-            SandboxTransition::Ready => {
-                self.phase = SandboxPhase::Ready;
-                self.error = None;
-                self.pause_snapshot_id = None;
-                self.paused_at = None;
-                self.redact_runtime_inputs();
-            }
-            SandboxTransition::Stopping => {
-                self.phase = SandboxPhase::Stopping;
-                self.error = None;
-                self.redact_runtime_inputs();
-            }
-            SandboxTransition::Stopped => {
-                self.phase = SandboxPhase::Stopped;
-                self.error = None;
-            }
-            SandboxTransition::Failed(error) => {
-                self.phase = SandboxPhase::Failed;
-                self.error = Some(error);
-                self.redact_runtime_inputs();
-            }
-            SandboxTransition::Removing => {
-                self.phase = SandboxPhase::Removing;
-                self.redact_runtime_inputs();
-            }
-            SandboxTransition::Pausing => {
-                self.phase = SandboxPhase::Pausing;
-                self.error = None;
-            }
-            SandboxTransition::Paused { snapshot_id } => {
-                self.phase = SandboxPhase::Paused;
-                self.pause_snapshot_id = Some(snapshot_id);
-                // Stamped once per pause, not once per transition: a failed
-                // resume parks the record back at `Paused`, and overwriting
-                // here would report the sandbox as freshly paused — and push
-                // the time forward again on every retry. `Ready` clears the
-                // stamp, so a genuine re-pause still gets a fresh one.
-                self.paused_at.get_or_insert_with(Utc::now);
-                self.error = None;
-            }
-            SandboxTransition::Resuming => {
-                self.phase = SandboxPhase::Resuming;
-                self.error = None;
-            }
-        }
-        Ok(())
-    }
-
-    fn redact_runtime_inputs(&mut self) {
-        self.effective_spec.kernel.clear();
-        self.effective_spec.rootfs.clear();
-        self.effective_spec.boot_args.clear();
-        self.effective_spec.cmd.clear();
-        self.effective_spec.env.clear();
-        self.effective_spec.working_dir.clear();
-        self.effective_spec.user.clear();
-        self.effective_spec.mounts.clear();
-        self.effective_spec.ttl_seconds = 0;
-        self.effective_spec.ssh_public_key = None;
-    }
-}
-
-impl SandboxPhase {
-    fn can_transition_to(self, next: Self) -> bool {
-        self == next
-            || matches!(
-                (self, next),
-                (
-                    Self::Creating,
-                    Self::Starting | Self::Failed | Self::Removing
-                ) | (
-                    Self::Starting,
-                    Self::Ready | Self::Stopping | Self::Failed | Self::Removing
-                ) | (
-                    Self::Ready,
-                    Self::Stopping | Self::Pausing | Self::Failed | Self::Removing
-                ) | (
-                    Self::Stopping,
-                    Self::Stopped | Self::Failed | Self::Removing
-                ) | (Self::Stopped | Self::Failed, Self::Removing)
-                    // A failed pause reverts to Ready and a completed one
-                    // parks at Paused; a resume mirrors that exactly (a
-                    // failed one unwinds back to Paused).
-                    | (
-                        Self::Pausing | Self::Resuming,
-                        Self::Paused | Self::Ready | Self::Failed | Self::Removing
-                    )
-                    | (Self::Paused, Self::Resuming | Self::Failed | Self::Removing)
-            )
-    }
-
-    pub(super) fn as_str(self) -> &'static str {
-        match self {
-            Self::Creating => "creating",
-            Self::Starting => "starting",
-            Self::Ready => "ready",
-            Self::Stopping => "stopping",
-            Self::Stopped => "stopped",
-            Self::Failed => "failed",
-            Self::Removing => "removing",
-            Self::Pausing => "pausing",
-            Self::Paused => "paused",
-            Self::Resuming => "resuming",
-        }
-    }
-}
 
 /// A filesystem mutation whose rename/unlink happened, but whose directory
 /// fsync may have failed. Callers must treat `value` as the visible state and
 /// may surface `durability_error` as an unconfirmed result.
-pub(super) struct DurableCommit<T> {
-    pub(super) value: T,
-    pub(super) durability_error: Option<String>,
+pub(in crate::sandbox) struct DurableCommit<T> {
+    pub(in crate::sandbox) value: T,
+    pub(in crate::sandbox) durability_error: Option<String>,
 }
 
 impl<T> DurableCommit<T> {
-    pub(super) fn confirmed(self, operation: &str) -> Result<T> {
+    pub(in crate::sandbox) fn confirmed(self, operation: &str) -> Result<T> {
         match self.durability_error {
             Some(error) => Err(VmmError::Unavailable(format!(
                 "{operation} is visible but its durability is unconfirmed: {error}"
@@ -307,21 +49,14 @@ impl<T> DurableCommit<T> {
 }
 
 /// File-backed sandbox record store serialized by a process-local mutex.
-pub(super) struct SandboxRecordStore {
+pub(in crate::sandbox) struct SandboxRecordStore {
     root: PathBuf,
     lock: Mutex<()>,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ExistingProvision {
-    Pending,
-    Replay,
-    Blocked,
-}
-
 impl SandboxRecordStore {
     /// Opens the durable store beside, rather than inside, runtime sandboxes.
-    pub(super) fn new(data_dir: &Path) -> Result<Self> {
+    pub(in crate::sandbox) fn new(data_dir: &Path) -> Result<Self> {
         let root = data_dir.join(RECORDS_DIR);
         fs::create_dir_all(&root)?;
         fs::set_permissions(&root, fs::Permissions::from_mode(0o700))?;
@@ -335,13 +70,13 @@ impl SandboxRecordStore {
 
     /// Loads the current record for `id`.
     #[cfg(test)]
-    pub(super) fn load(&self, id: &str) -> Result<Option<SandboxRecord>> {
+    pub(in crate::sandbox) fn load(&self, id: &str) -> Result<Option<SandboxRecord>> {
         let _guard = self.lock()?;
         self.load_unlocked(id)
     }
 
     /// Loads every durable record, rejecting any malformed record file.
-    pub(super) fn load_all(&self) -> Result<Vec<SandboxRecord>> {
+    pub(in crate::sandbox) fn load_all(&self) -> Result<Vec<SandboxRecord>> {
         let _guard = self.lock()?;
         let mut paths = fs::read_dir(&self.root)?
             .map(|entry| entry.map(|entry| entry.path()))
@@ -378,7 +113,7 @@ impl SandboxRecordStore {
     ///
     /// `None` means either no record exists or the matching create is still
     /// pending; both cases continue through the normal provisioning path.
-    pub(super) fn replay_provision(
+    pub(in crate::sandbox) fn replay_provision(
         &self,
         id: &str,
         request_key: &str,
@@ -403,7 +138,7 @@ impl SandboxRecordStore {
     }
 
     /// Persists a new provisioning intent or classifies the existing match.
-    pub(super) fn provision_intent(
+    pub(in crate::sandbox) fn provision_intent(
         &self,
         id: &str,
         request_key: &str,
@@ -437,7 +172,7 @@ impl SandboxRecordStore {
     }
 
     /// Applies a lifecycle transition only to the expected generation.
-    pub(super) fn transition(
+    pub(in crate::sandbox) fn transition(
         &self,
         id: &str,
         generation: Uuid,
@@ -463,7 +198,7 @@ impl SandboxRecordStore {
     /// Called by `SetLifecycle` (CORE-60) with the post-update values so a
     /// paused sandbox reloaded after an agent restart keeps its (re-armed)
     /// TTL deadline and idle policy.
-    pub(super) fn update_lifecycle(
+    pub(in crate::sandbox) fn update_lifecycle(
         &self,
         id: &str,
         generation: Uuid,
@@ -489,17 +224,28 @@ impl SandboxRecordStore {
     }
 
     /// Removes a known pre-ACK provision intent after side effects rolled back.
-    pub(super) fn abort_provision(&self, id: &str, generation: Uuid) -> Result<DurableCommit<()>> {
+    pub(in crate::sandbox) fn abort_provision(
+        &self,
+        id: &str,
+        generation: Uuid,
+    ) -> Result<DurableCommit<()>> {
         self.delete_record(id, generation, SandboxPhase::Creating)
     }
 
     /// Releases an ID only after removal and resource cleanup completed.
-    pub(super) fn finish_remove(&self, id: &str, generation: Uuid) -> Result<DurableCommit<()>> {
+    pub(in crate::sandbox) fn finish_remove(
+        &self,
+        id: &str,
+        generation: Uuid,
+    ) -> Result<DurableCommit<()>> {
         self.delete_record(id, generation, SandboxPhase::Removing)
     }
 
     /// Cancels a pre-ACK intent, or confirms that an already-removed ID is free.
-    pub(super) fn cancel_pending_or_missing(&self, id: &str) -> Result<DurableCommit<()>> {
+    pub(in crate::sandbox) fn cancel_pending_or_missing(
+        &self,
+        id: &str,
+    ) -> Result<DurableCommit<()>> {
         let _guard = self.lock()?;
         let Some(record) = self.load_unlocked(id)? else {
             return Ok(self.confirm_absence_unlocked());
@@ -637,84 +383,6 @@ fn validate_provision_input(id: &str, request_key: &str) -> Result<()> {
         return Err(VmmError::Config(
             "sandbox provision request key must not be empty".into(),
         ));
-    }
-    Ok(())
-}
-
-fn classify_existing_provision(
-    record: &SandboxRecord,
-    request_key: &str,
-) -> Result<ExistingProvision> {
-    if record.request_key != request_key {
-        return Err(VmmError::AlreadyExists(record.id.clone()));
-    }
-    Ok(match record.phase {
-        SandboxPhase::Creating => ExistingProvision::Pending,
-        SandboxPhase::Starting | SandboxPhase::Ready => ExistingProvision::Replay,
-        // A paused sandbox's provision outcome names a released IP, so a
-        // same-key create retry must not replay it as live.
-        SandboxPhase::Stopping
-        | SandboxPhase::Stopped
-        | SandboxPhase::Failed
-        | SandboxPhase::Removing
-        | SandboxPhase::Pausing
-        | SandboxPhase::Paused
-        | SandboxPhase::Resuming => ExistingProvision::Blocked,
-    })
-}
-
-fn validate_record(id: &str, record: &SandboxRecord) -> Result<()> {
-    validate_id("sandbox id", id)?;
-    if record.version != RECORD_VERSION {
-        return Err(VmmError::Config(format!(
-            "unsupported sandbox record version {} for {id}",
-            record.version
-        )));
-    }
-    if record.id != id {
-        return Err(VmmError::Config(format!(
-            "sandbox record id mismatch: expected {id}, got {}",
-            record.id
-        )));
-    }
-    if record.effective_spec.id.as_deref() != Some(id) {
-        return Err(VmmError::Config(format!(
-            "sandbox record spec id mismatch for {id}"
-        )));
-    }
-    if record.request_key.is_empty() {
-        return Err(VmmError::Config(format!(
-            "sandbox record provision request key is empty for {id}"
-        )));
-    }
-    if record.phase == SandboxPhase::Creating && record.provision_outcome.is_some() {
-        return Err(VmmError::Config(format!(
-            "creating sandbox record unexpectedly has a provision outcome for {id}"
-        )));
-    }
-    if matches!(
-        record.phase,
-        SandboxPhase::Starting
-            | SandboxPhase::Ready
-            | SandboxPhase::Stopping
-            | SandboxPhase::Stopped
-            | SandboxPhase::Pausing
-            | SandboxPhase::Paused
-            | SandboxPhase::Resuming
-    ) && record.provision_outcome.is_none()
-    {
-        return Err(VmmError::Config(format!(
-            "sandbox record has no provision outcome in phase {} for {id}",
-            record.phase.as_str()
-        )));
-    }
-    if matches!(record.phase, SandboxPhase::Paused | SandboxPhase::Resuming)
-        && record.pause_snapshot_id.is_none()
-    {
-        return Err(VmmError::Config(format!(
-            "sandbox record has no pause snapshot in phase {} for {id}",
-            record.phase.as_str()
-        )));
     }
     Ok(())
 }
